Externalize Dependency As PlanEntry
Previously the mapping of a BuildpackDependency to a BuildpackPlanEntry was
internal to a function.  However, as more complex layer contribution scenarios
have come about, the need to re-use that logic shows itself.  This change
moves the mapping behavior to a function on the BuildpackDependency so it can
be called by other components.

Signed-off-by: Ben Hale <bhale@pivotal.io>