DownloadCache Request Modification
This change adds an additional function to the DownloadCache to allow
modification of the outgoing request for a given artifact.  This would
typically be used to set Authorization headers on an outgoing request.

This change also adds an optional RequestModificationFunction to the
DependencyLayerHelper to enable users to modify the request that it creates.

Signed-off-by: Ben Hale <bhale@pivotal.io>