Limit FileEntry Concurrency
Previously, when listing the FileEntry's sha256 calculation would happen in an
uncontrolled manner.  Each calculation would open a file in a goroutine and
given a typical combination of disk speed and file count, the number of
simultaneously opened files could overwhelm the OS limit.  This change limits
the number of open files to 128 by managing a worker queue.

Signed-off-by: Ben Hale <bhale@pivotal.io>