Remove Modification Time
Previously, the FileListing would include the modification time of the files
in the listing.  Unfortunately, in systems like kpack where clones of Git
repositories happen in isolation from one another, the modification time is
always the time of the latest clone and therefore this metadata contained in
the listing will often be non-matching even when it otherwise should be.

This change removes modification time from the file listing metadata.

[resolves #2]

Signed-off-by: Ben Hale <bhale@vmware.com>