Cleanup Configuration Logging
Previously if a configuration value contained multiple lines (say if it came
from a nicely formatted XML configuration), those new lines would be reflected
in the output of the buildpack.  This both broke the tabular format and
generally looked bad.  This change collapse values so that they're always on a
single line.

Signed-off-by: Ben Hale <bhale@vmware.com>