Exec.d Updates
This change adds support for exec.d.  It does this in two parts.

First, the HelperLayer has been updated such that it no longer allows
arbitrary helper applications.  Instead it expects a single helper application
that will be symlinked to multiple filenames.  This will be considered the
optimal design for helpers in the future as it minimizes the size of the image.

Second, it introduces a new sherpa.Helper which is a wrapper around multiple
helper types and dispatches based on their names.  In addition it serializes
the returned map of environment variables to the exec.d output in fd/3.

[buildpacks/rfcs#104]

Signed-off-by: Ben Hale <bhale@vmware.com>