Support No package.json
Previously the NodeJSMainModule function would return an error if a NodeJS
application did not have a package.json file.  This was incorrect as it is
possible to have a legal NodeJS application without one and the Paketo NodeJS
buildpack supports this style of application.

This change updates the detection logic to return server.js in cases where
there is no package.json or the package.json does not explicitly identify a
main, in line with the behavior of the Paketo NodeJS buildpack.

Signed-off-by: Ben Hale <bhale@vmware.com>