Introduces a new interface for resolving bindings.
The previous BindingsResolver did not provide an interface for resolving multiple bindings of a given type. Some paketo use-cases such as dependency-mappings and ca-certficates should be able to support an arbitrary number of bindigns. This changes intoduces a new bindings package that exports both Resolve and ResolveOne. In addition it provides functionality that allows the consumer to filter on provider in addition to type.

Signed-off-by: Emily Casey <ecasey@vmware.com>