Skip `.git` folders in file listings
The file listing mechanism will traverse all directories given to it & return a listing of all files. The main purpose of this mechanism is to detect changes under a directory structure. This change will skip `.git` folders as the file listing is traversing subdirectories. This is because `.git` folders are well-know, can change independent of the application (like if you change git config) but shouldn't trigger rebuilds of the application. In additiona, the `.git` directory can for larger projects contain a large number of files, so skipping it can help with performance on large projects.

This change will skip all folders with the name `.git`, not just the top-level `.git` folder.

This resolves #55.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>