Do not assume buildpack id ends with `-buildpacks`
The package id regular expression assumes that the package ID will be `<something>-buildpacks/<id>`. This works for `paketo-buildpacks/` and `tanzu-buildpacks/`. It does not work for the `paketocommunity`.

This commit modifes the regular expression to permit anything in the form `<something>/<id>`, which matches previous patterns plus `paketocommunity/` as well.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>