Adds sherpa.ResolveBool for use in helpers
Provide the method with an env variable name and it will resolve & attempt to validate in a consistent way if the variable is true or false.

The function returns true for 1, t, T, TRUE, true, True. It returns false for all other values (even invalid) and if the variable is not set.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>