Fixes issues with incorrect indentation
Given a situation where bytes were written to the logger that do not end with a newline, then another set of bytes are written, the second set of bytes would be indented incorrectly. For example. First, write `some-data`, then write `more-data\nindent`. It should log `some-datamore-data\n    indent`, but it actually writes `some-data    more-data\n    indent`. There should not be an indent between the two writes because we only indent after a newline.

This caused issues when indenting output from commands like `mvn` and `cargo`.

This PR resolves the issue and adds tests to confirm correct output.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>