Removes logic to restrict output of deprecated label-based BOM entries
Starting with lifecycle 0.13.3, it is permitted to have both the old style label-based BOM information and the new style layer-based BOM information. If the buildpack API is 0.6 or older, label-based BOMs only is OK. If the buildpack API is 0.7, you may have both label-based BOM and layer-based BOM or just layer-based BOM. It is permitted to have just label-based BOM, however, that will generate a warning from the lifecycle.

Given this, we're removing the logic which conditionally controlled the BOM formats that were output. Going forward DependencyLayerContributor and HelperLayerContributor will always output both BOM formats.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>