Allow disabling the BOM label
We are currently supporting the old-style BOM label and the new-style SBOM layer. By default, we output both.

In some situations, if the BOM label grows large enough the resulting will not run on Kubernetes. It fails because the label is too large. You may now set the `BP_BOM_LABEL_DISABLED` env variable to `true` and on the next build, the BOM label will not be included with the image.

The label is included by default, despite this known issue, to for backward compatibility.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>