Add options to create-package for filtering dependencies
The create-package tool allows you to include dependencies with the buildpack that it packages. Prior to this commit, it was all or nothing. You'd get the full set of dependencies or none.

This PR provides two new flags to create-package which allow you to control what depenencies are included. The first is 'dependency-filter', which can be set multiple times and is a regex that indicates which dependencies are included. It by default will be matched against either the dependency ID or version. If either match the pattern, then they are included. The second option added is 'strict-filters', which is a boolean that if true means your filters must match both the dependency ID and version.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>