Extend libcnb's log.Logger as Paketo's log.Logger and implement
We add a new interface `log.Logger`, which has methods for all of the different log levels that we use in Paketo.

It also contains an implementation of PaketoLogger which implements Logger. The implementation uses color and formatting in the way that is consistent with libpak v1. In previous versions, debug and info level logging was done at the libcnb logging level. In libcnb v2, we can now control this and the PaketoLogger sets the debug logger to have cyan colored output. In addition, there is no Info level log. You can use Body in place of Info.

Other changes in this commit:

- NewConfigurationResolver no longer logs the configuration. You need to manually call `ConfigurationResolver.LogConfiguration` to log the configuration.
- ConfigurationResolver and DependencyResolver use the `log.Logger` which is an interface. If you have code that was depending on a default `log.Logger` struct, this will now break. With `log.Logger` being an interface, the default is nil. This may cause panics in your code.
- NewDependencyCache requires a logger. Previously, it didn't which was weird because it will log stuff. In general, pass it a logger but if you do not want it to log anything pass it `log.NewDiscardLogger()`
- NewDependencyLayerContributor requires a logger. Previously, it didn't which was weird because it will log stuff. In general, pass it a logger but if you do not want it to log anything pass it `log.NewDiscardLogger()`

Signed-off-by: Daniel Mikusa <dan@mikusa.com>