Add `ContributeLayersFunc` and reverse flow of `ContributableBuildFunc`
Previously, it was on the buildpack author to call `ContributableBuildFunc` which was a little awkward and created some testing issues. You could not mock out this method so it would run your layer contributors even during tests. In libpak v1, the layer contributor was returned from build and converted by libcnb into a layer so when doing tests, we could capture the layer contributors but not execute them.

This change flips the flow around so that `ContributableBuildFunc` takes a `ContributeLayersFunc` and the `ContributeLayersFunc` produces a list of layer contributors. Normally, `ContributableBuildFunc` would run and process the layer contributors making layers but in tests you can simply run the `ContributeLayersFunc` directly and check that the list of layer contributors is as expected making it more convenient for testing purposes.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>