Modify carton package to work with targets
We recently added targets to the `include-files` list in `buildpack.toml`. The referenced files get packaged up into the buildpack. What we added were some files prefixed with `linux/<arch>/...`. These files get put into that location within the generated buildpack directory. This is to work with a new version of pack that packages multi-arch images. It packages up whatever is under each architecture directory into the image specific to that architecture.

What's happened is that we now have non-architecture dependent files like the README and LICENSE which are not being included in the images. These files also need to be prefixed with the path for each architecture. Instead of putting a lot of duplicate entries into the `include-files` list, what this PR does is to take any file not prefixed with `linux/` and duplicate it for all of the detected architectures.

The PR determines the list of target architectures by looking at the `include-files` list and pulling out all of the distinct `linux/<arch>`'s that it sees. This isn't the best implementation but it's all we can do with the v1.x branch of libpak because the target metadata is not present in that version of libcnb. In libpak 2.x, we'll be able to read the target list directly from `buildpack.toml` and use that instead, which is a better option.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>