Remove commands that were moved to libpak-tools & move carton package into libpak-tools
The rest of this is moving the internal packages that were all used by libpak before to now be external packages so that they're accessible by libpak-tools. This is a trade-off that we'll need to make in order to split these up, but I think it's OK. Some of these can be useful outside of libpak.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>