Allow layer contributors to contribute process types
Resolves #349

This introduces a new interface `ProcessContributable` which extends `Contributable`. It introduces a new method that must be implemented `ProcessTypes`. This method will be called and allows you to return the list of process types that should be added.

The `ContributableBuildFunc` method is modified to check the type. If it is just `Contributable` then it adds a layer to the BuildResult. If it is `ProcessContributable` then it will add the layer and it will add the list of processes to the BuildResult too.

This should actually be a backward compatible way to implement this, which means existing Contributable implementations do not need to change. However, we do not have any existing Contributable implementations as this is a new interface in v2. At the same time, it should make migration to v2 easier as this is closer to the v1 implementation.

An alternative implementation for this would be to just have one interface `Contributable` and to add the `ProcessTypes` method to that. This would mean that for implementations that don't care to return process types, they would still need to implement the method though. A similar approach would be to modify the `Contribute` method on the `Contributable` interface to pass the process types back from the `Contribute` method. That would mean every implementation needs to deal with process types, but it could just return an empty list which is easy enough. The downside of that is then `Contributable` becomes more multi-purpose. I don't really have a stong preference on how we implement this, but went with the current implementation as it is flexible but it's separates concerns. If there's a strong preference for a different implemention, please let me know.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>